+++ src/test/kotlin/tornadofx/NodeTests.kt
@@ -1,11 +1,16 @@
 package tornadofx
 
+import javafx.application.Platform
+import javafx.scene.Parent
+import javafx.scene.control.Label
 import javafx.scene.control.TableColumn
+import javafx.scene.layout.HBox
 import javafx.scene.layout.StackPane
 import javafx.stage.Stage
 import org.junit.Before
 import org.junit.Test
 import org.testfx.api.FxToolkit
+import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
 class NodeTests {
@@ -24,4 +29,22 @@ class NodeTests {
         column.makeEditable()
         assertTrue(column.cellFactory != null)
     }
+
+    @Test
+    fun findComponents() {
+        val view1 = find(View1::class)
+        val view2 = find(View2::class)
+        val view3 = find(View3::class)
+
+        view1 += view2
+        view1 += view3
+
+        assertEquals(view2, view1.find<View2>())
+        assertEquals(view3, view1.find<View3>())
+        assertEquals(2, view1.findAll<View>().size)
+    }
+
+    class View1 : View() { override val root = HBox(Label("View 1")) }
+    class View2 : View() { override val root = HBox(Label("View 2")) }
+    class View3 : View() { override val root = HBox(Label("View 3")) }
 }